First half working version of the store.
